requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
sqlmodel==0.0.8
aiocache==0.11.1
redis==4.5.5
pydantic==1.10.12
msgpack==1.1.1
psycopg2-binary==2.9.10
pybase62==1.0.0